+++ ppb/systems/__init__.py
@@ -1,8 +1,7 @@
-from ppb.systems.base import System
 from ppb.systems.inputs import EventPoller
 from ppb.systems.renderer import Renderer, Image
 from ppb.systems.clocks import Updater
 
 __all__ = (
-    'System', 'EventPoller', 'Renderer', 'Image', 'Updater',
+    'EventPoller', 'Renderer', 'Image', 'Updater',
 )
